use canvas from chat-ui
