+++ packages/server/next/app/components/ui/chat/chat-message-content.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import { ChatMessage } from "@llamaindex/chat-ui";
-import { ArtifactCards } from "./custom/events/artifact-card";
 import { DynamicEvents } from "./custom/events/dynamic-events";
 import { ComponentDef } from "./custom/events/types";
 import { ToolAnnotations } from "./tools/chat-tools";
@@ -20,7 +19,7 @@ export function ChatMessageContent({
       <ToolAnnotations />
       <ChatMessage.Content.Image />
       <DynamicEvents componentDefs={componentDefs} appendError={appendError} />
-      <ArtifactCards />
+      <ChatMessage.Content.Artifact />
       <ChatMessage.Content.Markdown />
       <ChatMessage.Content.DocumentFile />
       <ChatMessage.Content.Source />
