+++ packages/server/next/app/components/ui/chat/chat-section.tsx
@@ -1,1745 +1,20 @@
 "use client";
 
-import { ChatSection as ChatUIContainer } from "@llamaindex/chat-ui";
-import { Message, useChat } from "ai/react";
+import { ChatCanvas, ChatSection as ChatUI } from "@llamaindex/chat-ui";
+import { useChat } from "ai/react";
 import { useEffect, useMemo, useState } from "react";
 import { getConfig } from "../lib/utils";
-import { ChatCanvas } from "./chat-canvas";
-import { ChatCanvasProvider } from "./chat-canvas-provider";
 import { ChatHeader } from "./chat-header";
 import { ChatInjection } from "./chat-injection";
 import CustomChatInput from "./chat-input";
 import CustomChatMessages from "./chat-messages";
+import { FrontendPreview } from "./custom/artifact-playground";
 import { DynamicEventsErrors } from "./custom/events/dynamic-events-errors";
 import { fetchComponentDefinitions } from "./custom/events/loader";
 import { ComponentDef } from "./custom/events/types";
 
-const initialMessages = [
-  {
-    role: "user",
-    content: "Generate a simple calculator in nextjs",
-    id: "hsQyTK5iOUr3I9TM",
-    createdAt: "2025-04-23T07:26:24.713Z",
-    parts: [
-      {
-        type: "text",
-        text: "Generate a simple calculator in nextjs",
-      },
-    ],
-  },
-  {
-    id: "6dbFXu7Ix8Nbe21K",
-    createdAt: "2025-04-23T07:26:24.730Z",
-    role: "assistant",
-    content:
-      'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-    parts: [
-      {
-        type: "text",
-        text: 'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-      },
-    ],
-    annotations: [
-      {
-        userInput: "Generate a simple calculator in nextjs",
-        chatHistory: [],
-      },
-      {
-        input: [
-          {
-            role: "user",
-            content: "Generate a simple calculator in nextjs",
-          },
-        ],
-        currentAgentName: "Agent",
-      },
-      {
-        input: [
-          {
-            role: "system",
-            content:
-              "You are a helpful assistant. Use the provided tools to answer questions.",
-          },
-          {
-            role: "user",
-            content: "Generate a simple calculator in nextjs",
-          },
-        ],
-        currentAgentName: "Agent",
-      },
-      {
-        agentName: "Agent",
-        response: {
-          role: "assistant",
-          content: "",
-          options: {
-            toolCall: [
-              {
-                name: "artifact_generator",
-                input: {
-                  requirement:
-                    "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                  artifactType: "code",
-                },
-                id: "call_MEP0izjICt3YrHQKw8lMm7re",
-              },
-            ],
-          },
-        },
-        toolCalls: [
-          {
-            data: {
-              agentName: "Agent",
-              toolName: "artifact_generator",
-              toolKwargs: {
-                requirement:
-                  "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                artifactType: "code",
-              },
-              toolId: "call_MEP0izjICt3YrHQKw8lMm7re",
-            },
-            displayName: "AgentToolCall",
-          },
-        ],
-      },
-      {
-        response: {
-          role: "assistant",
-          content: "",
-          options: {
-            toolCall: [
-              {
-                name: "artifact_generator",
-                input: {
-                  requirement:
-                    "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                  artifactType: "code",
-                },
-                id: "call_MEP0izjICt3YrHQKw8lMm7re",
-              },
-            ],
-          },
-        },
-        toolCalls: [
-          {
-            data: {
-              agentName: "Agent",
-              toolName: "artifact_generator",
-              toolKwargs: {
-                requirement:
-                  "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                artifactType: "code",
-              },
-              toolId: "call_MEP0izjICt3YrHQKw8lMm7re",
-            },
-            displayName: "AgentToolCall",
-          },
-        ],
-        currentAgentName: "Agent",
-      },
-      {
-        agentName: "Agent",
-        toolCalls: [
-          {
-            data: {
-              agentName: "Agent",
-              toolName: "artifact_generator",
-              toolKwargs: {
-                requirement:
-                  "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                artifactType: "code",
-              },
-              toolId: "call_MEP0izjICt3YrHQKw8lMm7re",
-            },
-            displayName: "AgentToolCall",
-          },
-        ],
-      },
-      {
-        type: "agent",
-        data: {
-          agent: "Agent",
-          text: 'Using tool: \'artifact_generator\' with inputs: \'{"requirement":"Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.","artifactType":"code"}\'',
-          type: "text",
-        },
-      },
-      {
-        agentName: "Agent",
-        results: [
-          {
-            data: {
-              toolName: "artifact_generator",
-              toolKwargs: {
-                requirement:
-                  "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                artifactType: "code",
-              },
-              toolId: "call_MEP0izjICt3YrHQKw8lMm7re",
-              toolOutput: {
-                id: "call_MEP0izjICt3YrHQKw8lMm7re",
-                result:
-                  "{\n  type: code,\n  data: {\n    file_name: Calculator.tsx,\n    code: import React, { useState } from 'react';\\n\\nconst Calculator: React.FC = () => {\\n  const [input, setInput] = useState<string>('');\\n  const [result, setResult] = useState<number | null>(null);\\n\\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\\n    setInput(e.target.value);\\n  };\\n\\n  const calculateResult = () => {\\n    try {\\n      // Evaluate the expression using Function constructor\\n      const evalResult = new Function('return ' + input)();\\n      setResult(evalResult);\\n    } catch (error) {\\n      setResult(null);\\n      alert('Invalid expression');\\n    }\\n  };\\n\\n  return (\\n    <div className=\\p-4 max-w-md mx-auto\\>\\n      <h1 className=\\text-2xl font-bold mb-4\\>Simple Calculator</h1>\\n      <input\\n        type=\\text\\\\n        value={input}\\n        onChange={handleInputChange}\\n        className=\\border p-2 w-full mb-4\\\\n        placeholder=\\Enter expression (e.g., 2+2)\\\\n      />\\n      <button\\n        onClick={calculateResult}\\n        className=\\bg-blue-500 text-white p-2 w-full mb-4\\\\n      >\\n        Calculate\\n      </button>\\n      {result !== null && (\\n        <div className=\\text-xl\\>Result: {result}</div>\\n      )}\\n    </div>\\n  );\\n};\\n\\nexport default Calculator;,\n    language: typescript\n  },\n  created_at: 1745393189834\n}",
-                isError: false,
-              },
-              returnDirect: false,
-              raw: {
-                type: "code",
-                data: {
-                  file_name: "Calculator.tsx",
-                  code: 'import React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;',
-                  language: "typescript",
-                },
-                created_at: 1745393189834,
-              },
-            },
-            displayName: "AgentToolCallResult",
-          },
-        ],
-      },
-      {
-        type: "artifact",
-        data: {
-          type: "code",
-          data: {
-            file_name: "Calculator.tsx",
-            code: 'import React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;',
-            language: "typescript",
-          },
-          created_at: 1745393189834,
-        },
-      },
-      {
-        input: [
-          {
-            role: "user",
-            content: "Generate a simple calculator in nextjs",
-          },
-        ],
-        currentAgentName: "Agent",
-      },
-      {
-        input: [
-          {
-            role: "system",
-            content:
-              "You are a helpful assistant. Use the provided tools to answer questions.",
-          },
-          {
-            role: "user",
-            content: "Generate a simple calculator in nextjs",
-          },
-        ],
-        currentAgentName: "Agent",
-      },
-      {
-        agentName: "Agent",
-        response: {
-          role: "assistant",
-          content:
-            'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-        },
-        toolCalls: [],
-      },
-      {
-        response: {
-          role: "assistant",
-          content:
-            'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-        },
-        toolCalls: [],
-        currentAgentName: "Agent",
-      },
-      {
-        result:
-          'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-      },
-      {
-        type: "suggested_questions",
-        data: [
-          "How can I add more advanced operations like square root or exponentiation to this calculator?",
-          "Is it safe to use the Function constructor for evaluating expressions in a web application?",
-          "Can you show me how to style this calculator using CSS modules or styled-components?",
-        ],
-      },
-    ],
-    revisionId: "sDuxP0O9NhXSOtKb",
-  },
-  {
-    role: "user",
-    content: "Make it has numbers from 1-9, AC button",
-    annotations: [],
-    id: "EVQb7PjrQU6CVsZl",
-    createdAt: "2025-04-23T07:27:03.330Z",
-    parts: [
-      {
-        type: "text",
-        text: "Make it has numbers from 1-9, AC button",
-      },
-    ],
-  },
-  {
-    id: "W5eQ7winNgX2jKnK",
-    createdAt: "2025-04-23T07:27:03.354Z",
-    role: "assistant",
-    content:
-      "Here is an updated version of the simple calculator in Next.js that includes buttons for numbers 1-9, an \"AC\" (All Clear) button, and basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from 'react';\nimport { Button } from 'shadcn-ui';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState('');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleButtonClick = (value: string) => {\n    setInput((prev) => prev + value);\n  };\n\n  const handleClear = () => {\n    setInput('');\n    setResult(null);\n  };\n\n  const handleCalculate = () => {\n    try {\n      // Evaluate the expression\n      const evalResult = eval(input);\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert('Invalid expression');\n    }\n  };\n\n  return (\n    <div className=\"max-w-xs mx-auto mt-10 p-4 border rounded shadow-lg\">\n      <div className=\"mb-4 text-right\">\n        <div className=\"text-lg font-mono\">{input || '0'}</div>\n        <div className=\"text-2xl font-bold\">{result !== null ? result : ''}</div>\n      </div>\n      <div className=\"grid grid-cols-4 gap-2\">\n        {[1, 2, 3, 4, 5, 6, 7, 8, 9].map((num) => (\n          <Button key={num} onClick={() => handleButtonClick(num.toString())}>\n            {num}\n          </Button>\n        ))}\n        <Button onClick={handleClear} className=\"col-span-2\">\n          AC\n        </Button>\n        <Button onClick={() => handleButtonClick('+')}>+</Button>\n        <Button onClick={() => handleButtonClick('-')}>-</Button>\n        <Button onClick={() => handleButtonClick('*')}>*</Button>\n        <Button onClick={() => handleButtonClick('/')}>/</Button>\n        <Button onClick={handleCalculate} className=\"col-span-4\">\n          =\n        </Button>\n      </div>\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Ensure you have a Next.js project set up.\n2. Create a new file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Make sure to install any necessary UI library (like `shadcn-ui`) if you want to use the `Button` component as shown.\n5. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input numbers from 1 to 9, perform basic arithmetic operations, and clear the input with the \"AC\" button.",
-    parts: [
-      {
-        type: "text",
-        text: "Here is an updated version of the simple calculator in Next.js that includes buttons for numbers 1-9, an \"AC\" (All Clear) button, and basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from 'react';\nimport { Button } from 'shadcn-ui';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState('');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleButtonClick = (value: string) => {\n    setInput((prev) => prev + value);\n  };\n\n  const handleClear = () => {\n    setInput('');\n    setResult(null);\n  };\n\n  const handleCalculate = () => {\n    try {\n      // Evaluate the expression\n      const evalResult = eval(input);\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert('Invalid expression');\n    }\n  };\n\n  return (\n    <div className=\"max-w-xs mx-auto mt-10 p-4 border rounded shadow-lg\">\n      <div className=\"mb-4 text-right\">\n        <div className=\"text-lg font-mono\">{input || '0'}</div>\n        <div className=\"text-2xl font-bold\">{result !== null ? result : ''}</div>\n      </div>\n      <div className=\"grid grid-cols-4 gap-2\">\n        {[1, 2, 3, 4, 5, 6, 7, 8, 9].map((num) => (\n          <Button key={num} onClick={() => handleButtonClick(num.toString())}>\n            {num}\n          </Button>\n        ))}\n        <Button onClick={handleClear} className=\"col-span-2\">\n          AC\n        </Button>\n        <Button onClick={() => handleButtonClick('+')}>+</Button>\n        <Button onClick={() => handleButtonClick('-')}>-</Button>\n        <Button onClick={() => handleButtonClick('*')}>*</Button>\n        <Button onClick={() => handleButtonClick('/')}>/</Button>\n        <Button onClick={handleCalculate} className=\"col-span-4\">\n          =\n        </Button>\n      </div>\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Ensure you have a Next.js project set up.\n2. Create a new file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Make sure to install any necessary UI library (like `shadcn-ui`) if you want to use the `Button` component as shown.\n5. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input numbers from 1 to 9, perform basic arithmetic operations, and clear the input with the \"AC\" button.",
-      },
-    ],
-    annotations: [
-      {
-        userInput: "Make it has numbers from 1-9, AC button",
-        chatHistory: [
-          {
-            role: "user",
-            content: "Generate a simple calculator in nextjs",
-            parts: [
-              {
-                type: "text",
-                text: "Generate a simple calculator in nextjs",
-              },
-            ],
-          },
-          {
-            role: "assistant",
-            content:
-              'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-            annotations: [
-              {
-                userInput: "Generate a simple calculator in nextjs",
-                chatHistory: [],
-              },
-              {
-                input: [
-                  {
-                    role: "user",
-                    content: "Generate a simple calculator in nextjs",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                input: [
-                  {
-                    role: "system",
-                    content:
-                      "You are a helpful assistant. Use the provided tools to answer questions.",
-                  },
-                  {
-                    role: "user",
-                    content: "Generate a simple calculator in nextjs",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                agentName: "Agent",
-                response: {
-                  role: "assistant",
-                  content: "",
-                  options: {
-                    toolCall: [
-                      {
-                        name: "artifact_generator",
-                        input: {
-                          requirement:
-                            "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                          artifactType: "code",
-                        },
-                        id: "call_MEP0izjICt3YrHQKw8lMm7re",
-                      },
-                    ],
-                  },
-                },
-                toolCalls: [
-                  {
-                    data: {
-                      agentName: "Agent",
-                      toolName: "artifact_generator",
-                      toolKwargs: {
-                        requirement:
-                          "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                        artifactType: "code",
-                      },
-                      toolId: "call_MEP0izjICt3YrHQKw8lMm7re",
-                    },
-                    displayName: "AgentToolCall",
-                  },
-                ],
-              },
-              {
-                response: {
-                  role: "assistant",
-                  content: "",
-                  options: {
-                    toolCall: [
-                      {
-                        name: "artifact_generator",
-                        input: {
-                          requirement:
-                            "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                          artifactType: "code",
-                        },
-                        id: "call_MEP0izjICt3YrHQKw8lMm7re",
-                      },
-                    ],
-                  },
-                },
-                toolCalls: [
-                  {
-                    data: {
-                      agentName: "Agent",
-                      toolName: "artifact_generator",
-                      toolKwargs: {
-                        requirement:
-                          "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                        artifactType: "code",
-                      },
-                      toolId: "call_MEP0izjICt3YrHQKw8lMm7re",
-                    },
-                    displayName: "AgentToolCall",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                agentName: "Agent",
-                toolCalls: [
-                  {
-                    data: {
-                      agentName: "Agent",
-                      toolName: "artifact_generator",
-                      toolKwargs: {
-                        requirement:
-                          "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                        artifactType: "code",
-                      },
-                      toolId: "call_MEP0izjICt3YrHQKw8lMm7re",
-                    },
-                    displayName: "AgentToolCall",
-                  },
-                ],
-              },
-              {
-                type: "agent",
-                data: {
-                  agent: "Agent",
-                  text: 'Using tool: \'artifact_generator\' with inputs: \'{"requirement":"Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.","artifactType":"code"}\'',
-                  type: "text",
-                },
-              },
-              {
-                agentName: "Agent",
-                results: [
-                  {
-                    data: {
-                      toolName: "artifact_generator",
-                      toolKwargs: {
-                        requirement:
-                          "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                        artifactType: "code",
-                      },
-                      toolId: "call_MEP0izjICt3YrHQKw8lMm7re",
-                      toolOutput: {
-                        id: "call_MEP0izjICt3YrHQKw8lMm7re",
-                        result:
-                          "{\n  type: code,\n  data: {\n    file_name: Calculator.tsx,\n    code: import React, { useState } from 'react';\\n\\nconst Calculator: React.FC = () => {\\n  const [input, setInput] = useState<string>('');\\n  const [result, setResult] = useState<number | null>(null);\\n\\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\\n    setInput(e.target.value);\\n  };\\n\\n  const calculateResult = () => {\\n    try {\\n      // Evaluate the expression using Function constructor\\n      const evalResult = new Function('return ' + input)();\\n      setResult(evalResult);\\n    } catch (error) {\\n      setResult(null);\\n      alert('Invalid expression');\\n    }\\n  };\\n\\n  return (\\n    <div className=\\p-4 max-w-md mx-auto\\>\\n      <h1 className=\\text-2xl font-bold mb-4\\>Simple Calculator</h1>\\n      <input\\n        type=\\text\\\\n        value={input}\\n        onChange={handleInputChange}\\n        className=\\border p-2 w-full mb-4\\\\n        placeholder=\\Enter expression (e.g., 2+2)\\\\n      />\\n      <button\\n        onClick={calculateResult}\\n        className=\\bg-blue-500 text-white p-2 w-full mb-4\\\\n      >\\n        Calculate\\n      </button>\\n      {result !== null && (\\n        <div className=\\text-xl\\>Result: {result}</div>\\n      )}\\n    </div>\\n  );\\n};\\n\\nexport default Calculator;,\n    language: typescript\n  },\n  created_at: 1745393189834\n}",
-                        isError: false,
-                      },
-                      returnDirect: false,
-                      raw: {
-                        type: "code",
-                        data: {
-                          file_name: "Calculator.tsx",
-                          code: 'import React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;',
-                          language: "typescript",
-                        },
-                        created_at: 1745393189834,
-                      },
-                    },
-                    displayName: "AgentToolCallResult",
-                  },
-                ],
-              },
-              {
-                type: "artifact",
-                data: {
-                  type: "code",
-                  data: {
-                    file_name: "Calculator.tsx",
-                    code: 'import React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;',
-                    language: "typescript",
-                  },
-                  created_at: 1745393189834,
-                },
-              },
-              {
-                input: [
-                  {
-                    role: "user",
-                    content: "Generate a simple calculator in nextjs",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                input: [
-                  {
-                    role: "system",
-                    content:
-                      "You are a helpful assistant. Use the provided tools to answer questions.",
-                  },
-                  {
-                    role: "user",
-                    content: "Generate a simple calculator in nextjs",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                agentName: "Agent",
-                response: {
-                  role: "assistant",
-                  content:
-                    'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-                },
-                toolCalls: [],
-              },
-              {
-                response: {
-                  role: "assistant",
-                  content:
-                    'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-                },
-                toolCalls: [],
-                currentAgentName: "Agent",
-              },
-              {
-                result:
-                  'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-              },
-              {
-                type: "suggested_questions",
-                data: [
-                  "How can I add more advanced operations like square root or exponentiation to this calculator?",
-                  "Is it safe to use the Function constructor for evaluating expressions in a web application?",
-                  "Can you show me how to style this calculator using CSS modules or styled-components?",
-                ],
-              },
-            ],
-            parts: [
-              {
-                type: "text",
-                text: 'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-              },
-            ],
-          },
-        ],
-      },
-      {
-        input: [
-          {
-            role: "user",
-            content: "Generate a simple calculator in nextjs",
-            parts: [
-              {
-                type: "text",
-                text: "Generate a simple calculator in nextjs",
-              },
-            ],
-          },
-          {
-            role: "assistant",
-            content:
-              'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-            annotations: [
-              {
-                userInput: "Generate a simple calculator in nextjs",
-                chatHistory: [],
-              },
-              {
-                input: [
-                  {
-                    role: "user",
-                    content: "Generate a simple calculator in nextjs",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                input: [
-                  {
-                    role: "system",
-                    content:
-                      "You are a helpful assistant. Use the provided tools to answer questions.",
-                  },
-                  {
-                    role: "user",
-                    content: "Generate a simple calculator in nextjs",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                agentName: "Agent",
-                response: {
-                  role: "assistant",
-                  content: "",
-                  options: {
-                    toolCall: [
-                      {
-                        name: "artifact_generator",
-                        input: {
-                          requirement:
-                            "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                          artifactType: "code",
-                        },
-                        id: "call_MEP0izjICt3YrHQKw8lMm7re",
-                      },
-                    ],
-                  },
-                },
-                toolCalls: [
-                  {
-                    data: {
-                      agentName: "Agent",
-                      toolName: "artifact_generator",
-                      toolKwargs: {
-                        requirement:
-                          "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                        artifactType: "code",
-                      },
-                      toolId: "call_MEP0izjICt3YrHQKw8lMm7re",
-                    },
-                    displayName: "AgentToolCall",
-                  },
-                ],
-              },
-              {
-                response: {
-                  role: "assistant",
-                  content: "",
-                  options: {
-                    toolCall: [
-                      {
-                        name: "artifact_generator",
-                        input: {
-                          requirement:
-                            "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                          artifactType: "code",
-                        },
-                        id: "call_MEP0izjICt3YrHQKw8lMm7re",
-                      },
-                    ],
-                  },
-                },
-                toolCalls: [
-                  {
-                    data: {
-                      agentName: "Agent",
-                      toolName: "artifact_generator",
-                      toolKwargs: {
-                        requirement:
-                          "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                        artifactType: "code",
-                      },
-                      toolId: "call_MEP0izjICt3YrHQKw8lMm7re",
-                    },
-                    displayName: "AgentToolCall",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                agentName: "Agent",
-                toolCalls: [
-                  {
-                    data: {
-                      agentName: "Agent",
-                      toolName: "artifact_generator",
-                      toolKwargs: {
-                        requirement:
-                          "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                        artifactType: "code",
-                      },
-                      toolId: "call_MEP0izjICt3YrHQKw8lMm7re",
-                    },
-                    displayName: "AgentToolCall",
-                  },
-                ],
-              },
-              {
-                type: "agent",
-                data: {
-                  agent: "Agent",
-                  text: 'Using tool: \'artifact_generator\' with inputs: \'{"requirement":"Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.","artifactType":"code"}\'',
-                  type: "text",
-                },
-              },
-              {
-                agentName: "Agent",
-                results: [
-                  {
-                    data: {
-                      toolName: "artifact_generator",
-                      toolKwargs: {
-                        requirement:
-                          "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                        artifactType: "code",
-                      },
-                      toolId: "call_MEP0izjICt3YrHQKw8lMm7re",
-                      toolOutput: {
-                        id: "call_MEP0izjICt3YrHQKw8lMm7re",
-                        result:
-                          "{\n  type: code,\n  data: {\n    file_name: Calculator.tsx,\n    code: import React, { useState } from 'react';\\n\\nconst Calculator: React.FC = () => {\\n  const [input, setInput] = useState<string>('');\\n  const [result, setResult] = useState<number | null>(null);\\n\\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\\n    setInput(e.target.value);\\n  };\\n\\n  const calculateResult = () => {\\n    try {\\n      // Evaluate the expression using Function constructor\\n      const evalResult = new Function('return ' + input)();\\n      setResult(evalResult);\\n    } catch (error) {\\n      setResult(null);\\n      alert('Invalid expression');\\n    }\\n  };\\n\\n  return (\\n    <div className=\\p-4 max-w-md mx-auto\\>\\n      <h1 className=\\text-2xl font-bold mb-4\\>Simple Calculator</h1>\\n      <input\\n        type=\\text\\\\n        value={input}\\n        onChange={handleInputChange}\\n        className=\\border p-2 w-full mb-4\\\\n        placeholder=\\Enter expression (e.g., 2+2)\\\\n      />\\n      <button\\n        onClick={calculateResult}\\n        className=\\bg-blue-500 text-white p-2 w-full mb-4\\\\n      >\\n        Calculate\\n      </button>\\n      {result !== null && (\\n        <div className=\\text-xl\\>Result: {result}</div>\\n      )}\\n    </div>\\n  );\\n};\\n\\nexport default Calculator;,\n    language: typescript\n  },\n  created_at: 1745393189834\n}",
-                        isError: false,
-                      },
-                      returnDirect: false,
-                      raw: {
-                        type: "code",
-                        data: {
-                          file_name: "Calculator.tsx",
-                          code: 'import React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;',
-                          language: "typescript",
-                        },
-                        created_at: 1745393189834,
-                      },
-                    },
-                    displayName: "AgentToolCallResult",
-                  },
-                ],
-              },
-              {
-                type: "artifact",
-                data: {
-                  type: "code",
-                  data: {
-                    file_name: "Calculator.tsx",
-                    code: 'import React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;',
-                    language: "typescript",
-                  },
-                  created_at: 1745393189834,
-                },
-              },
-              {
-                input: [
-                  {
-                    role: "user",
-                    content: "Generate a simple calculator in nextjs",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                input: [
-                  {
-                    role: "system",
-                    content:
-                      "You are a helpful assistant. Use the provided tools to answer questions.",
-                  },
-                  {
-                    role: "user",
-                    content: "Generate a simple calculator in nextjs",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                agentName: "Agent",
-                response: {
-                  role: "assistant",
-                  content:
-                    'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-                },
-                toolCalls: [],
-              },
-              {
-                response: {
-                  role: "assistant",
-                  content:
-                    'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-                },
-                toolCalls: [],
-                currentAgentName: "Agent",
-              },
-              {
-                result:
-                  'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-              },
-              {
-                type: "suggested_questions",
-                data: [
-                  "How can I add more advanced operations like square root or exponentiation to this calculator?",
-                  "Is it safe to use the Function constructor for evaluating expressions in a web application?",
-                  "Can you show me how to style this calculator using CSS modules or styled-components?",
-                ],
-              },
-            ],
-            parts: [
-              {
-                type: "text",
-                text: 'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-              },
-            ],
-          },
-          {
-            role: "user",
-            content: "Make it has numbers from 1-9, AC button",
-          },
-        ],
-        currentAgentName: "Agent",
-      },
-      {
-        input: [
-          {
-            role: "system",
-            content:
-              "You are a helpful assistant. Use the provided tools to answer questions.",
-          },
-          {
-            role: "user",
-            content: "Generate a simple calculator in nextjs",
-            parts: [
-              {
-                type: "text",
-                text: "Generate a simple calculator in nextjs",
-              },
-            ],
-          },
-          {
-            role: "assistant",
-            content:
-              'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-            annotations: [
-              {
-                userInput: "Generate a simple calculator in nextjs",
-                chatHistory: [],
-              },
-              {
-                input: [
-                  {
-                    role: "user",
-                    content: "Generate a simple calculator in nextjs",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                input: [
-                  {
-                    role: "system",
-                    content:
-                      "You are a helpful assistant. Use the provided tools to answer questions.",
-                  },
-                  {
-                    role: "user",
-                    content: "Generate a simple calculator in nextjs",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                agentName: "Agent",
-                response: {
-                  role: "assistant",
-                  content: "",
-                  options: {
-                    toolCall: [
-                      {
-                        name: "artifact_generator",
-                        input: {
-                          requirement:
-                            "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                          artifactType: "code",
-                        },
-                        id: "call_MEP0izjICt3YrHQKw8lMm7re",
-                      },
-                    ],
-                  },
-                },
-                toolCalls: [
-                  {
-                    data: {
-                      agentName: "Agent",
-                      toolName: "artifact_generator",
-                      toolKwargs: {
-                        requirement:
-                          "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                        artifactType: "code",
-                      },
-                      toolId: "call_MEP0izjICt3YrHQKw8lMm7re",
-                    },
-                    displayName: "AgentToolCall",
-                  },
-                ],
-              },
-              {
-                response: {
-                  role: "assistant",
-                  content: "",
-                  options: {
-                    toolCall: [
-                      {
-                        name: "artifact_generator",
-                        input: {
-                          requirement:
-                            "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                          artifactType: "code",
-                        },
-                        id: "call_MEP0izjICt3YrHQKw8lMm7re",
-                      },
-                    ],
-                  },
-                },
-                toolCalls: [
-                  {
-                    data: {
-                      agentName: "Agent",
-                      toolName: "artifact_generator",
-                      toolKwargs: {
-                        requirement:
-                          "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                        artifactType: "code",
-                      },
-                      toolId: "call_MEP0izjICt3YrHQKw8lMm7re",
-                    },
-                    displayName: "AgentToolCall",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                agentName: "Agent",
-                toolCalls: [
-                  {
-                    data: {
-                      agentName: "Agent",
-                      toolName: "artifact_generator",
-                      toolKwargs: {
-                        requirement:
-                          "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                        artifactType: "code",
-                      },
-                      toolId: "call_MEP0izjICt3YrHQKw8lMm7re",
-                    },
-                    displayName: "AgentToolCall",
-                  },
-                ],
-              },
-              {
-                type: "agent",
-                data: {
-                  agent: "Agent",
-                  text: 'Using tool: \'artifact_generator\' with inputs: \'{"requirement":"Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.","artifactType":"code"}\'',
-                  type: "text",
-                },
-              },
-              {
-                agentName: "Agent",
-                results: [
-                  {
-                    data: {
-                      toolName: "artifact_generator",
-                      toolKwargs: {
-                        requirement:
-                          "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                        artifactType: "code",
-                      },
-                      toolId: "call_MEP0izjICt3YrHQKw8lMm7re",
-                      toolOutput: {
-                        id: "call_MEP0izjICt3YrHQKw8lMm7re",
-                        result:
-                          "{\n  type: code,\n  data: {\n    file_name: Calculator.tsx,\n    code: import React, { useState } from 'react';\\n\\nconst Calculator: React.FC = () => {\\n  const [input, setInput] = useState<string>('');\\n  const [result, setResult] = useState<number | null>(null);\\n\\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\\n    setInput(e.target.value);\\n  };\\n\\n  const calculateResult = () => {\\n    try {\\n      // Evaluate the expression using Function constructor\\n      const evalResult = new Function('return ' + input)();\\n      setResult(evalResult);\\n    } catch (error) {\\n      setResult(null);\\n      alert('Invalid expression');\\n    }\\n  };\\n\\n  return (\\n    <div className=\\p-4 max-w-md mx-auto\\>\\n      <h1 className=\\text-2xl font-bold mb-4\\>Simple Calculator</h1>\\n      <input\\n        type=\\text\\\\n        value={input}\\n        onChange={handleInputChange}\\n        className=\\border p-2 w-full mb-4\\\\n        placeholder=\\Enter expression (e.g., 2+2)\\\\n      />\\n      <button\\n        onClick={calculateResult}\\n        className=\\bg-blue-500 text-white p-2 w-full mb-4\\\\n      >\\n        Calculate\\n      </button>\\n      {result !== null && (\\n        <div className=\\text-xl\\>Result: {result}</div>\\n      )}\\n    </div>\\n  );\\n};\\n\\nexport default Calculator;,\n    language: typescript\n  },\n  created_at: 1745393189834\n}",
-                        isError: false,
-                      },
-                      returnDirect: false,
-                      raw: {
-                        type: "code",
-                        data: {
-                          file_name: "Calculator.tsx",
-                          code: 'import React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;',
-                          language: "typescript",
-                        },
-                        created_at: 1745393189834,
-                      },
-                    },
-                    displayName: "AgentToolCallResult",
-                  },
-                ],
-              },
-              {
-                type: "artifact",
-                data: {
-                  type: "code",
-                  data: {
-                    file_name: "Calculator.tsx",
-                    code: 'import React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;',
-                    language: "typescript",
-                  },
-                  created_at: 1745393189834,
-                },
-              },
-              {
-                input: [
-                  {
-                    role: "user",
-                    content: "Generate a simple calculator in nextjs",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                input: [
-                  {
-                    role: "system",
-                    content:
-                      "You are a helpful assistant. Use the provided tools to answer questions.",
-                  },
-                  {
-                    role: "user",
-                    content: "Generate a simple calculator in nextjs",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                agentName: "Agent",
-                response: {
-                  role: "assistant",
-                  content:
-                    'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-                },
-                toolCalls: [],
-              },
-              {
-                response: {
-                  role: "assistant",
-                  content:
-                    'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-                },
-                toolCalls: [],
-                currentAgentName: "Agent",
-              },
-              {
-                result:
-                  'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-              },
-              {
-                type: "suggested_questions",
-                data: [
-                  "How can I add more advanced operations like square root or exponentiation to this calculator?",
-                  "Is it safe to use the Function constructor for evaluating expressions in a web application?",
-                  "Can you show me how to style this calculator using CSS modules or styled-components?",
-                ],
-              },
-            ],
-            parts: [
-              {
-                type: "text",
-                text: 'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-              },
-            ],
-          },
-          {
-            role: "user",
-            content: "Make it has numbers from 1-9, AC button",
-          },
-        ],
-        currentAgentName: "Agent",
-      },
-      {
-        agentName: "Agent",
-        response: {
-          role: "assistant",
-          content: "",
-          options: {
-            toolCall: [
-              {
-                name: "artifact_generator",
-                input: {
-                  requirement:
-                    "Create a simple calculator in Next.js that includes buttons for numbers 1-9, an 'AC' (All Clear) button, and basic arithmetic operations. The calculator should display the input and result clearly.",
-                  artifactType: "code",
-                },
-                id: "call_aKTH8zbZMtsGTsGaqAcsu1Is",
-              },
-            ],
-          },
-        },
-        toolCalls: [
-          {
-            data: {
-              agentName: "Agent",
-              toolName: "artifact_generator",
-              toolKwargs: {
-                requirement:
-                  "Create a simple calculator in Next.js that includes buttons for numbers 1-9, an 'AC' (All Clear) button, and basic arithmetic operations. The calculator should display the input and result clearly.",
-                artifactType: "code",
-              },
-              toolId: "call_aKTH8zbZMtsGTsGaqAcsu1Is",
-            },
-            displayName: "AgentToolCall",
-          },
-        ],
-      },
-      {
-        response: {
-          role: "assistant",
-          content: "",
-          options: {
-            toolCall: [
-              {
-                name: "artifact_generator",
-                input: {
-                  requirement:
-                    "Create a simple calculator in Next.js that includes buttons for numbers 1-9, an 'AC' (All Clear) button, and basic arithmetic operations. The calculator should display the input and result clearly.",
-                  artifactType: "code",
-                },
-                id: "call_aKTH8zbZMtsGTsGaqAcsu1Is",
-              },
-            ],
-          },
-        },
-        toolCalls: [
-          {
-            data: {
-              agentName: "Agent",
-              toolName: "artifact_generator",
-              toolKwargs: {
-                requirement:
-                  "Create a simple calculator in Next.js that includes buttons for numbers 1-9, an 'AC' (All Clear) button, and basic arithmetic operations. The calculator should display the input and result clearly.",
-                artifactType: "code",
-              },
-              toolId: "call_aKTH8zbZMtsGTsGaqAcsu1Is",
-            },
-            displayName: "AgentToolCall",
-          },
-        ],
-        currentAgentName: "Agent",
-      },
-      {
-        agentName: "Agent",
-        toolCalls: [
-          {
-            data: {
-              agentName: "Agent",
-              toolName: "artifact_generator",
-              toolKwargs: {
-                requirement:
-                  "Create a simple calculator in Next.js that includes buttons for numbers 1-9, an 'AC' (All Clear) button, and basic arithmetic operations. The calculator should display the input and result clearly.",
-                artifactType: "code",
-              },
-              toolId: "call_aKTH8zbZMtsGTsGaqAcsu1Is",
-            },
-            displayName: "AgentToolCall",
-          },
-        ],
-      },
-      {
-        type: "agent",
-        data: {
-          agent: "Agent",
-          text: 'Using tool: \'artifact_generator\' with inputs: \'{"requirement":"Create a simple calculator in Next.js that includes buttons for numbers 1-9, an \'AC\' (All Clear) button, and basic arithmetic operations. The calculator should display the input and result clearly.","artifactType":"code"}\'',
-          type: "text",
-        },
-      },
-      {
-        agentName: "Agent",
-        results: [
-          {
-            data: {
-              toolName: "artifact_generator",
-              toolKwargs: {
-                requirement:
-                  "Create a simple calculator in Next.js that includes buttons for numbers 1-9, an 'AC' (All Clear) button, and basic arithmetic operations. The calculator should display the input and result clearly.",
-                artifactType: "code",
-              },
-              toolId: "call_aKTH8zbZMtsGTsGaqAcsu1Is",
-              toolOutput: {
-                id: "call_aKTH8zbZMtsGTsGaqAcsu1Is",
-                result:
-                  "{\n  type: code,\n  data: {\n    file_name: Calculator.tsx,\n    code: import React, { useState } from 'react';\\nimport { Button } from 'shadcn-ui';\\n\\nconst Calculator: React.FC = () => {\\n  const [input, setInput] = useState('');\\n  const [result, setResult] = useState<number | null>(null);\\n\\n  const handleButtonClick = (value: string) => {\\n    setInput((prev) => prev + value);\\n  };\\n\\n  const handleClear = () => {\\n    setInput('');\\n    setResult(null);\\n  };\\n\\n  const handleCalculate = () => {\\n    try {\\n      // Evaluate the expression\\n      const evalResult = eval(input);\\n      setResult(evalResult);\\n    } catch (error) {\\n      setResult(null);\\n      alert('Invalid expression');\\n    }\\n  };\\n\\n  return (\\n    <div className=\\max-w-xs mx-auto mt-10 p-4 border rounded shadow-lg\\>\\n      <div className=\\mb-4 text-right\\>\\n        <div className=\\text-lg font-mono\\>{input || '0'}</div>\\n        <div className=\\text-2xl font-bold\\>{result !== null ? result : ''}</div>\\n      </div>\\n      <div className=\\grid grid-cols-4 gap-2\\>\\n        {[1, 2, 3, 4, 5, 6, 7, 8, 9].map((num) => (\\n          <Button key={num} onClick={() => handleButtonClick(num.toString())}>\\n            {num}\\n          </Button>\\n        ))}\\n        <Button onClick={handleClear} className=\\col-span-2\\>\\n          AC\\n        </Button>\\n        <Button onClick={() => handleButtonClick('+')}>+</Button>\\n        <Button onClick={() => handleButtonClick('-')}>-</Button>\\n        <Button onClick={() => handleButtonClick('*')}>*</Button>\\n        <Button onClick={() => handleButtonClick('/')}>/</Button>\\n        <Button onClick={handleCalculate} className=\\col-span-4\\>\\n          =\\n        </Button>\\n      </div>\\n    </div>\\n  );\\n};\\n\\nexport default Calculator;,\n    language: typescript\n  },\n  created_at: 1745393231347\n}",
-                isError: false,
-              },
-              returnDirect: false,
-              raw: {
-                type: "code",
-                data: {
-                  file_name: "Calculator.tsx",
-                  code: "import React, { useState } from 'react';\nimport { Button } from 'shadcn-ui';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState('');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleButtonClick = (value: string) => {\n    setInput((prev) => prev + value);\n  };\n\n  const handleClear = () => {\n    setInput('');\n    setResult(null);\n  };\n\n  const handleCalculate = () => {\n    try {\n      // Evaluate the expression\n      const evalResult = eval(input);\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert('Invalid expression');\n    }\n  };\n\n  return (\n    <div className=\"max-w-xs mx-auto mt-10 p-4 border rounded shadow-lg\">\n      <div className=\"mb-4 text-right\">\n        <div className=\"text-lg font-mono\">{input || '0'}</div>\n        <div className=\"text-2xl font-bold\">{result !== null ? result : ''}</div>\n      </div>\n      <div className=\"grid grid-cols-4 gap-2\">\n        {[1, 2, 3, 4, 5, 6, 7, 8, 9].map((num) => (\n          <Button key={num} onClick={() => handleButtonClick(num.toString())}>\n            {num}\n          </Button>\n        ))}\n        <Button onClick={handleClear} className=\"col-span-2\">\n          AC\n        </Button>\n        <Button onClick={() => handleButtonClick('+')}>+</Button>\n        <Button onClick={() => handleButtonClick('-')}>-</Button>\n        <Button onClick={() => handleButtonClick('*')}>*</Button>\n        <Button onClick={() => handleButtonClick('/')}>/</Button>\n        <Button onClick={handleCalculate} className=\"col-span-4\">\n          =\n        </Button>\n      </div>\n    </div>\n  );\n};\n\nexport default Calculator;",
-                  language: "typescript",
-                },
-                created_at: 1745393231347,
-              },
-            },
-            displayName: "AgentToolCallResult",
-          },
-        ],
-      },
-      {
-        type: "artifact",
-        data: {
-          type: "code",
-          data: {
-            file_name: "Calculator.tsx",
-            code: "import React, { useState } from 'react';\nimport { Button } from 'shadcn-ui';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState('');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleButtonClick = (value: string) => {\n    setInput((prev) => prev + value);\n  };\n\n  const handleClear = () => {\n    setInput('');\n    setResult(null);\n  };\n\n  const handleCalculate = () => {\n    try {\n      // Evaluate the expression\n      const evalResult = eval(input);\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert('Invalid expression');\n    }\n  };\n\n  return (\n    <div className=\"max-w-xs mx-auto mt-10 p-4 border rounded shadow-lg\">\n      <div className=\"mb-4 text-right\">\n        <div className=\"text-lg font-mono\">{input || '0'}</div>\n        <div className=\"text-2xl font-bold\">{result !== null ? result : ''}</div>\n      </div>\n      <div className=\"grid grid-cols-4 gap-2\">\n        {[1, 2, 3, 4, 5, 6, 7, 8, 9].map((num) => (\n          <Button key={num} onClick={() => handleButtonClick(num.toString())}>\n            {num}\n          </Button>\n        ))}\n        <Button onClick={handleClear} className=\"col-span-2\">\n          AC\n        </Button>\n        <Button onClick={() => handleButtonClick('+')}>+</Button>\n        <Button onClick={() => handleButtonClick('-')}>-</Button>\n        <Button onClick={() => handleButtonClick('*')}>*</Button>\n        <Button onClick={() => handleButtonClick('/')}>/</Button>\n        <Button onClick={handleCalculate} className=\"col-span-4\">\n          =\n        </Button>\n      </div>\n    </div>\n  );\n};\n\nexport default Calculator;",
-            language: "typescript",
-          },
-          created_at: 1745393231347,
-        },
-      },
-      {
-        input: [
-          {
-            role: "user",
-            content: "Generate a simple calculator in nextjs",
-            parts: [
-              {
-                type: "text",
-                text: "Generate a simple calculator in nextjs",
-              },
-            ],
-          },
-          {
-            role: "assistant",
-            content:
-              'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-            annotations: [
-              {
-                userInput: "Generate a simple calculator in nextjs",
-                chatHistory: [],
-              },
-              {
-                input: [
-                  {
-                    role: "user",
-                    content: "Generate a simple calculator in nextjs",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                input: [
-                  {
-                    role: "system",
-                    content:
-                      "You are a helpful assistant. Use the provided tools to answer questions.",
-                  },
-                  {
-                    role: "user",
-                    content: "Generate a simple calculator in nextjs",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                agentName: "Agent",
-                response: {
-                  role: "assistant",
-                  content: "",
-                  options: {
-                    toolCall: [
-                      {
-                        name: "artifact_generator",
-                        input: {
-                          requirement:
-                            "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                          artifactType: "code",
-                        },
-                        id: "call_MEP0izjICt3YrHQKw8lMm7re",
-                      },
-                    ],
-                  },
-                },
-                toolCalls: [
-                  {
-                    data: {
-                      agentName: "Agent",
-                      toolName: "artifact_generator",
-                      toolKwargs: {
-                        requirement:
-                          "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                        artifactType: "code",
-                      },
-                      toolId: "call_MEP0izjICt3YrHQKw8lMm7re",
-                    },
-                    displayName: "AgentToolCall",
-                  },
-                ],
-              },
-              {
-                response: {
-                  role: "assistant",
-                  content: "",
-                  options: {
-                    toolCall: [
-                      {
-                        name: "artifact_generator",
-                        input: {
-                          requirement:
-                            "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                          artifactType: "code",
-                        },
-                        id: "call_MEP0izjICt3YrHQKw8lMm7re",
-                      },
-                    ],
-                  },
-                },
-                toolCalls: [
-                  {
-                    data: {
-                      agentName: "Agent",
-                      toolName: "artifact_generator",
-                      toolKwargs: {
-                        requirement:
-                          "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                        artifactType: "code",
-                      },
-                      toolId: "call_MEP0izjICt3YrHQKw8lMm7re",
-                    },
-                    displayName: "AgentToolCall",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                agentName: "Agent",
-                toolCalls: [
-                  {
-                    data: {
-                      agentName: "Agent",
-                      toolName: "artifact_generator",
-                      toolKwargs: {
-                        requirement:
-                          "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                        artifactType: "code",
-                      },
-                      toolId: "call_MEP0izjICt3YrHQKw8lMm7re",
-                    },
-                    displayName: "AgentToolCall",
-                  },
-                ],
-              },
-              {
-                type: "agent",
-                data: {
-                  agent: "Agent",
-                  text: 'Using tool: \'artifact_generator\' with inputs: \'{"requirement":"Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.","artifactType":"code"}\'',
-                  type: "text",
-                },
-              },
-              {
-                agentName: "Agent",
-                results: [
-                  {
-                    data: {
-                      toolName: "artifact_generator",
-                      toolKwargs: {
-                        requirement:
-                          "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                        artifactType: "code",
-                      },
-                      toolId: "call_MEP0izjICt3YrHQKw8lMm7re",
-                      toolOutput: {
-                        id: "call_MEP0izjICt3YrHQKw8lMm7re",
-                        result:
-                          "{\n  type: code,\n  data: {\n    file_name: Calculator.tsx,\n    code: import React, { useState } from 'react';\\n\\nconst Calculator: React.FC = () => {\\n  const [input, setInput] = useState<string>('');\\n  const [result, setResult] = useState<number | null>(null);\\n\\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\\n    setInput(e.target.value);\\n  };\\n\\n  const calculateResult = () => {\\n    try {\\n      // Evaluate the expression using Function constructor\\n      const evalResult = new Function('return ' + input)();\\n      setResult(evalResult);\\n    } catch (error) {\\n      setResult(null);\\n      alert('Invalid expression');\\n    }\\n  };\\n\\n  return (\\n    <div className=\\p-4 max-w-md mx-auto\\>\\n      <h1 className=\\text-2xl font-bold mb-4\\>Simple Calculator</h1>\\n      <input\\n        type=\\text\\\\n        value={input}\\n        onChange={handleInputChange}\\n        className=\\border p-2 w-full mb-4\\\\n        placeholder=\\Enter expression (e.g., 2+2)\\\\n      />\\n      <button\\n        onClick={calculateResult}\\n        className=\\bg-blue-500 text-white p-2 w-full mb-4\\\\n      >\\n        Calculate\\n      </button>\\n      {result !== null && (\\n        <div className=\\text-xl\\>Result: {result}</div>\\n      )}\\n    </div>\\n  );\\n};\\n\\nexport default Calculator;,\n    language: typescript\n  },\n  created_at: 1745393189834\n}",
-                        isError: false,
-                      },
-                      returnDirect: false,
-                      raw: {
-                        type: "code",
-                        data: {
-                          file_name: "Calculator.tsx",
-                          code: 'import React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;',
-                          language: "typescript",
-                        },
-                        created_at: 1745393189834,
-                      },
-                    },
-                    displayName: "AgentToolCallResult",
-                  },
-                ],
-              },
-              {
-                type: "artifact",
-                data: {
-                  type: "code",
-                  data: {
-                    file_name: "Calculator.tsx",
-                    code: 'import React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;',
-                    language: "typescript",
-                  },
-                  created_at: 1745393189834,
-                },
-              },
-              {
-                input: [
-                  {
-                    role: "user",
-                    content: "Generate a simple calculator in nextjs",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                input: [
-                  {
-                    role: "system",
-                    content:
-                      "You are a helpful assistant. Use the provided tools to answer questions.",
-                  },
-                  {
-                    role: "user",
-                    content: "Generate a simple calculator in nextjs",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                agentName: "Agent",
-                response: {
-                  role: "assistant",
-                  content:
-                    'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-                },
-                toolCalls: [],
-              },
-              {
-                response: {
-                  role: "assistant",
-                  content:
-                    'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-                },
-                toolCalls: [],
-                currentAgentName: "Agent",
-              },
-              {
-                result:
-                  'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-              },
-              {
-                type: "suggested_questions",
-                data: [
-                  "How can I add more advanced operations like square root or exponentiation to this calculator?",
-                  "Is it safe to use the Function constructor for evaluating expressions in a web application?",
-                  "Can you show me how to style this calculator using CSS modules or styled-components?",
-                ],
-              },
-            ],
-            parts: [
-              {
-                type: "text",
-                text: 'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-              },
-            ],
-          },
-          {
-            role: "user",
-            content: "Make it has numbers from 1-9, AC button",
-          },
-        ],
-        currentAgentName: "Agent",
-      },
-      {
-        input: [
-          {
-            role: "system",
-            content:
-              "You are a helpful assistant. Use the provided tools to answer questions.",
-          },
-          {
-            role: "user",
-            content: "Generate a simple calculator in nextjs",
-            parts: [
-              {
-                type: "text",
-                text: "Generate a simple calculator in nextjs",
-              },
-            ],
-          },
-          {
-            role: "assistant",
-            content:
-              'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-            annotations: [
-              {
-                userInput: "Generate a simple calculator in nextjs",
-                chatHistory: [],
-              },
-              {
-                input: [
-                  {
-                    role: "user",
-                    content: "Generate a simple calculator in nextjs",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                input: [
-                  {
-                    role: "system",
-                    content:
-                      "You are a helpful assistant. Use the provided tools to answer questions.",
-                  },
-                  {
-                    role: "user",
-                    content: "Generate a simple calculator in nextjs",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                agentName: "Agent",
-                response: {
-                  role: "assistant",
-                  content: "",
-                  options: {
-                    toolCall: [
-                      {
-                        name: "artifact_generator",
-                        input: {
-                          requirement:
-                            "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                          artifactType: "code",
-                        },
-                        id: "call_MEP0izjICt3YrHQKw8lMm7re",
-                      },
-                    ],
-                  },
-                },
-                toolCalls: [
-                  {
-                    data: {
-                      agentName: "Agent",
-                      toolName: "artifact_generator",
-                      toolKwargs: {
-                        requirement:
-                          "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                        artifactType: "code",
-                      },
-                      toolId: "call_MEP0izjICt3YrHQKw8lMm7re",
-                    },
-                    displayName: "AgentToolCall",
-                  },
-                ],
-              },
-              {
-                response: {
-                  role: "assistant",
-                  content: "",
-                  options: {
-                    toolCall: [
-                      {
-                        name: "artifact_generator",
-                        input: {
-                          requirement:
-                            "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                          artifactType: "code",
-                        },
-                        id: "call_MEP0izjICt3YrHQKw8lMm7re",
-                      },
-                    ],
-                  },
-                },
-                toolCalls: [
-                  {
-                    data: {
-                      agentName: "Agent",
-                      toolName: "artifact_generator",
-                      toolKwargs: {
-                        requirement:
-                          "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                        artifactType: "code",
-                      },
-                      toolId: "call_MEP0izjICt3YrHQKw8lMm7re",
-                    },
-                    displayName: "AgentToolCall",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                agentName: "Agent",
-                toolCalls: [
-                  {
-                    data: {
-                      agentName: "Agent",
-                      toolName: "artifact_generator",
-                      toolKwargs: {
-                        requirement:
-                          "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                        artifactType: "code",
-                      },
-                      toolId: "call_MEP0izjICt3YrHQKw8lMm7re",
-                    },
-                    displayName: "AgentToolCall",
-                  },
-                ],
-              },
-              {
-                type: "agent",
-                data: {
-                  agent: "Agent",
-                  text: 'Using tool: \'artifact_generator\' with inputs: \'{"requirement":"Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.","artifactType":"code"}\'',
-                  type: "text",
-                },
-              },
-              {
-                agentName: "Agent",
-                results: [
-                  {
-                    data: {
-                      toolName: "artifact_generator",
-                      toolKwargs: {
-                        requirement:
-                          "Create a simple calculator application using Next.js that can perform basic arithmetic operations like addition, subtraction, multiplication, and division.",
-                        artifactType: "code",
-                      },
-                      toolId: "call_MEP0izjICt3YrHQKw8lMm7re",
-                      toolOutput: {
-                        id: "call_MEP0izjICt3YrHQKw8lMm7re",
-                        result:
-                          "{\n  type: code,\n  data: {\n    file_name: Calculator.tsx,\n    code: import React, { useState } from 'react';\\n\\nconst Calculator: React.FC = () => {\\n  const [input, setInput] = useState<string>('');\\n  const [result, setResult] = useState<number | null>(null);\\n\\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\\n    setInput(e.target.value);\\n  };\\n\\n  const calculateResult = () => {\\n    try {\\n      // Evaluate the expression using Function constructor\\n      const evalResult = new Function('return ' + input)();\\n      setResult(evalResult);\\n    } catch (error) {\\n      setResult(null);\\n      alert('Invalid expression');\\n    }\\n  };\\n\\n  return (\\n    <div className=\\p-4 max-w-md mx-auto\\>\\n      <h1 className=\\text-2xl font-bold mb-4\\>Simple Calculator</h1>\\n      <input\\n        type=\\text\\\\n        value={input}\\n        onChange={handleInputChange}\\n        className=\\border p-2 w-full mb-4\\\\n        placeholder=\\Enter expression (e.g., 2+2)\\\\n      />\\n      <button\\n        onClick={calculateResult}\\n        className=\\bg-blue-500 text-white p-2 w-full mb-4\\\\n      >\\n        Calculate\\n      </button>\\n      {result !== null && (\\n        <div className=\\text-xl\\>Result: {result}</div>\\n      )}\\n    </div>\\n  );\\n};\\n\\nexport default Calculator;,\n    language: typescript\n  },\n  created_at: 1745393189834\n}",
-                        isError: false,
-                      },
-                      returnDirect: false,
-                      raw: {
-                        type: "code",
-                        data: {
-                          file_name: "Calculator.tsx",
-                          code: 'import React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;',
-                          language: "typescript",
-                        },
-                        created_at: 1745393189834,
-                      },
-                    },
-                    displayName: "AgentToolCallResult",
-                  },
-                ],
-              },
-              {
-                type: "artifact",
-                data: {
-                  type: "code",
-                  data: {
-                    file_name: "Calculator.tsx",
-                    code: 'import React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;',
-                    language: "typescript",
-                  },
-                  created_at: 1745393189834,
-                },
-              },
-              {
-                input: [
-                  {
-                    role: "user",
-                    content: "Generate a simple calculator in nextjs",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                input: [
-                  {
-                    role: "system",
-                    content:
-                      "You are a helpful assistant. Use the provided tools to answer questions.",
-                  },
-                  {
-                    role: "user",
-                    content: "Generate a simple calculator in nextjs",
-                  },
-                ],
-                currentAgentName: "Agent",
-              },
-              {
-                agentName: "Agent",
-                response: {
-                  role: "assistant",
-                  content:
-                    'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-                },
-                toolCalls: [],
-              },
-              {
-                response: {
-                  role: "assistant",
-                  content:
-                    'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-                },
-                toolCalls: [],
-                currentAgentName: "Agent",
-              },
-              {
-                result:
-                  'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-              },
-              {
-                type: "suggested_questions",
-                data: [
-                  "How can I add more advanced operations like square root or exponentiation to this calculator?",
-                  "Is it safe to use the Function constructor for evaluating expressions in a web application?",
-                  "Can you show me how to style this calculator using CSS modules or styled-components?",
-                ],
-              },
-            ],
-            parts: [
-              {
-                type: "text",
-                text: 'Here is a simple calculator application created using Next.js. The code is written in TypeScript and allows users to perform basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from \'react\';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState<string>(\'\');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {\n    setInput(e.target.value);\n  };\n\n  const calculateResult = () => {\n    try {\n      // Evaluate the expression using Function constructor\n      const evalResult = new Function(\'return \' + input)();\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert(\'Invalid expression\');\n    }\n  };\n\n  return (\n    <div className="p-4 max-w-md mx-auto">\n      <h1 className="text-2xl font-bold mb-4">Simple Calculator</h1>\n      <input\n        type="text"\n        value={input}\n        onChange={handleInputChange}\n        className="border p-2 w-full mb-4"\n        placeholder="Enter expression (e.g., 2+2)"\n      />\n      <button\n        onClick={calculateResult}\n        className="bg-blue-500 text-white p-2 w-full mb-4"\n      >\n        Calculate\n      </button>\n      {result !== null && (\n        <div className="text-xl">Result: {result}</div>\n      )}\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Create a new Next.js project if you haven\'t already.\n2. Add the above code in a file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input expressions like `2+2` and will display the result when the "Calculate" button is clicked.',
-              },
-            ],
-          },
-          {
-            role: "user",
-            content: "Make it has numbers from 1-9, AC button",
-          },
-        ],
-        currentAgentName: "Agent",
-      },
-      {
-        agentName: "Agent",
-        response: {
-          role: "assistant",
-          content:
-            "Here is an updated version of the simple calculator in Next.js that includes buttons for numbers 1-9, an \"AC\" (All Clear) button, and basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from 'react';\nimport { Button } from 'shadcn-ui';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState('');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleButtonClick = (value: string) => {\n    setInput((prev) => prev + value);\n  };\n\n  const handleClear = () => {\n    setInput('');\n    setResult(null);\n  };\n\n  const handleCalculate = () => {\n    try {\n      // Evaluate the expression\n      const evalResult = eval(input);\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert('Invalid expression');\n    }\n  };\n\n  return (\n    <div className=\"max-w-xs mx-auto mt-10 p-4 border rounded shadow-lg\">\n      <div className=\"mb-4 text-right\">\n        <div className=\"text-lg font-mono\">{input || '0'}</div>\n        <div className=\"text-2xl font-bold\">{result !== null ? result : ''}</div>\n      </div>\n      <div className=\"grid grid-cols-4 gap-2\">\n        {[1, 2, 3, 4, 5, 6, 7, 8, 9].map((num) => (\n          <Button key={num} onClick={() => handleButtonClick(num.toString())}>\n            {num}\n          </Button>\n        ))}\n        <Button onClick={handleClear} className=\"col-span-2\">\n          AC\n        </Button>\n        <Button onClick={() => handleButtonClick('+')}>+</Button>\n        <Button onClick={() => handleButtonClick('-')}>-</Button>\n        <Button onClick={() => handleButtonClick('*')}>*</Button>\n        <Button onClick={() => handleButtonClick('/')}>/</Button>\n        <Button onClick={handleCalculate} className=\"col-span-4\">\n          =\n        </Button>\n      </div>\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Ensure you have a Next.js project set up.\n2. Create a new file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Make sure to install any necessary UI library (like `shadcn-ui`) if you want to use the `Button` component as shown.\n5. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input numbers from 1 to 9, perform basic arithmetic operations, and clear the input with the \"AC\" button.",
-        },
-        toolCalls: [],
-      },
-      {
-        response: {
-          role: "assistant",
-          content:
-            "Here is an updated version of the simple calculator in Next.js that includes buttons for numbers 1-9, an \"AC\" (All Clear) button, and basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from 'react';\nimport { Button } from 'shadcn-ui';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState('');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleButtonClick = (value: string) => {\n    setInput((prev) => prev + value);\n  };\n\n  const handleClear = () => {\n    setInput('');\n    setResult(null);\n  };\n\n  const handleCalculate = () => {\n    try {\n      // Evaluate the expression\n      const evalResult = eval(input);\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert('Invalid expression');\n    }\n  };\n\n  return (\n    <div className=\"max-w-xs mx-auto mt-10 p-4 border rounded shadow-lg\">\n      <div className=\"mb-4 text-right\">\n        <div className=\"text-lg font-mono\">{input || '0'}</div>\n        <div className=\"text-2xl font-bold\">{result !== null ? result : ''}</div>\n      </div>\n      <div className=\"grid grid-cols-4 gap-2\">\n        {[1, 2, 3, 4, 5, 6, 7, 8, 9].map((num) => (\n          <Button key={num} onClick={() => handleButtonClick(num.toString())}>\n            {num}\n          </Button>\n        ))}\n        <Button onClick={handleClear} className=\"col-span-2\">\n          AC\n        </Button>\n        <Button onClick={() => handleButtonClick('+')}>+</Button>\n        <Button onClick={() => handleButtonClick('-')}>-</Button>\n        <Button onClick={() => handleButtonClick('*')}>*</Button>\n        <Button onClick={() => handleButtonClick('/')}>/</Button>\n        <Button onClick={handleCalculate} className=\"col-span-4\">\n          =\n        </Button>\n      </div>\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Ensure you have a Next.js project set up.\n2. Create a new file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Make sure to install any necessary UI library (like `shadcn-ui`) if you want to use the `Button` component as shown.\n5. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input numbers from 1 to 9, perform basic arithmetic operations, and clear the input with the \"AC\" button.",
-        },
-        toolCalls: [],
-        currentAgentName: "Agent",
-      },
-      {
-        result:
-          "Here is an updated version of the simple calculator in Next.js that includes buttons for numbers 1-9, an \"AC\" (All Clear) button, and basic arithmetic operations.\n\n### File: `Calculator.tsx`\n\n```typescript\nimport React, { useState } from 'react';\nimport { Button } from 'shadcn-ui';\n\nconst Calculator: React.FC = () => {\n  const [input, setInput] = useState('');\n  const [result, setResult] = useState<number | null>(null);\n\n  const handleButtonClick = (value: string) => {\n    setInput((prev) => prev + value);\n  };\n\n  const handleClear = () => {\n    setInput('');\n    setResult(null);\n  };\n\n  const handleCalculate = () => {\n    try {\n      // Evaluate the expression\n      const evalResult = eval(input);\n      setResult(evalResult);\n    } catch (error) {\n      setResult(null);\n      alert('Invalid expression');\n    }\n  };\n\n  return (\n    <div className=\"max-w-xs mx-auto mt-10 p-4 border rounded shadow-lg\">\n      <div className=\"mb-4 text-right\">\n        <div className=\"text-lg font-mono\">{input || '0'}</div>\n        <div className=\"text-2xl font-bold\">{result !== null ? result : ''}</div>\n      </div>\n      <div className=\"grid grid-cols-4 gap-2\">\n        {[1, 2, 3, 4, 5, 6, 7, 8, 9].map((num) => (\n          <Button key={num} onClick={() => handleButtonClick(num.toString())}>\n            {num}\n          </Button>\n        ))}\n        <Button onClick={handleClear} className=\"col-span-2\">\n          AC\n        </Button>\n        <Button onClick={() => handleButtonClick('+')}>+</Button>\n        <Button onClick={() => handleButtonClick('-')}>-</Button>\n        <Button onClick={() => handleButtonClick('*')}>*</Button>\n        <Button onClick={() => handleButtonClick('/')}>/</Button>\n        <Button onClick={handleCalculate} className=\"col-span-4\">\n          =\n        </Button>\n      </div>\n    </div>\n  );\n};\n\nexport default Calculator;\n```\n\n### Instructions to Use:\n1. Ensure you have a Next.js project set up.\n2. Create a new file named `Calculator.tsx` in your components directory.\n3. Import and use the `Calculator` component in your main application file (e.g., `pages/index.tsx`).\n4. Make sure to install any necessary UI library (like `shadcn-ui`) if you want to use the `Button` component as shown.\n5. Run your Next.js application to see the calculator in action.\n\nThis calculator allows users to input numbers from 1 to 9, perform basic arithmetic operations, and clear the input with the \"AC\" button.",
-      },
-      {
-        type: "suggested_questions",
-        data: [
-          "How can I add a decimal point button to the calculator?",
-          "Can you show me how to style the calculator using CSS modules in Next.js?",
-          "What changes do I need to make to include a history of calculations?",
-        ],
-      },
-    ],
-    revisionId: "vQOY4np4gYy9IIaL",
-  },
-];
-
 export default function ChatSection() {
   const handler = useChat({
-    initialMessages: initialMessages as unknown as Message[],
     api: getConfig("CHAT_API"),
     onError: (error: unknown) => {
       if (!(error instanceof Error)) throw error;
@@ -1757,24 +32,20 @@ export default function ChatSection() {
     <>
       <div className="flex h-screen w-screen flex-col overflow-hidden">
         <ChatHeader />
-        <ChatUIContainer
+        <ChatUI
           handler={handler}
           className="flex min-h-0 flex-1 flex-row justify-center gap-4 px-4 py-0"
         >
-          <ChatCanvasProvider
-            addMessages={(messages) => {
-              // append messages to chat history without triggering the API call
-              // TODO: we should support `setMessages` in chat-ui provider later then this props can be removed
-              handler.setMessages((prev) => [
-                ...prev,
-                ...(messages as Message[]),
-              ]);
-            }}
-          >
-            <CustomChatSection />
-            <ChatCanvas />
-          </ChatCanvasProvider>
-        </ChatUIContainer>
+          <CustomChatSection />
+          <ChatCanvas>
+            <ChatCanvas.CodeArtifact
+              tabs={{
+                preview: <FrontendPreview />,
+              }}
+            />
+            <ChatCanvas.DocumentArtifact />
+          </ChatCanvas>
+        </ChatUI>
       </div>
       <ChatInjection />
     </>
